# examples/client/bodies/api.yaml
openapi: 3.1.0
info: {title: request bodies, version: "1"}
paths:
...
  /form:
    post:
      operationId: postForm
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded: {schema: {$ref: '#/components/schemas/Note'}}
      responses:
        "200":
          description: echo
          content:
            application/json: {schema: {$ref: '#/components/schemas/Note'}}
    get:
      operationId: getForm
      responses:
        "200":
          description: a note as a form
          content:
            application/x-www-form-urlencoded: {schema: {$ref: '#/components/schemas/Note'}}
...
components:
  schemas:
    Note:
      type: object
      required: [text]
      properties:
        text: {type: string}
        stars: {type: integer}