# examples/server/beego/params/gen.go
// Code generated by mockzilla-codegen. DO NOT EDIT.

package params

import (
	"context"
	"errors"
	"io"
	"net/http"
	"strings"
	"time"

	"github.com/beego/beego/v2/server/web"
	bcontext "github.com/beego/beego/v2/server/web/context"
	"github.com/mockzilla/mockzilla-codegen/pkg/runtime"
)

// Fails to compile when the runtime does not match the generator that wrote this file.
const _ = runtime.SupportsGeneratorV2

type Point struct {
	X *int `json:"x,omitempty"`
	Y *int `json:"y,omitempty"`
}

type Echo map[string]any

type PathStylesPathParams struct {
	Simple string   `json:"simple"`
	Label  int      `json:"label"`
	Matrix bool     `json:"matrix"`
	List   []string `json:"list"`
}

// Validate checks the value against the constraints of the spec.
func (p PathStylesPathParams) Validate() error {
	var errs runtime.ValidationErrors
	if p.List == nil {
		errs.Add("list", "is required")
	}
	return errs.Err()
}

type QueryStylesQuery struct {
	Form   []int    `json:"form,omitempty"`
	Csv    []string `json:"csv,omitempty"`
	Space  []string `json:"space,omitempty"`
	Pipe   []string `json:"pipe,omitempty"`
	Deep   *Point   `json:"deep,omitempty"`
	Flat   *Point   `json:"flat,omitempty"`
	JSON   *Point   `json:"json,omitempty"`
	Needed string   `json:"needed"`
}

type HeaderStylesHeaders struct {
	XTags  []string   `json:"X-Tags,omitempty"`
	XPoint *Point     `json:"X-Point,omitempty"`
	XWhen  *time.Time `json:"X-When,omitempty"`
}

type CookieStylesCookies struct {
	Session *string `json:"session,omitempty"`
	Flags   []int   `json:"flags,omitempty"`
}

// ServiceInterface is what the generated handlers call. Implement it with the business logic.
type ServiceInterface interface {
	// PathStyles handles GET /path/{simple}/{label}/{matrix}/{list}.
	PathStyles(ctx context.Context, opts *PathStylesServiceRequestOptions) (*PathStylesResponseData, error)
	// QueryStyles handles GET /query.
	QueryStyles(ctx context.Context, opts *QueryStylesServiceRequestOptions) (*QueryStylesResponseData, error)
	// HeaderStyles handles GET /header.
	HeaderStyles(ctx context.Context, opts *HeaderStylesServiceRequestOptions) (*HeaderStylesResponseData, error)
	// CookieStyles handles GET /cookie.
	CookieStyles(ctx context.Context, opts *CookieStylesServiceRequestOptions) (*CookieStylesResponseData, error)
}

// PathStylesServiceRequestOptions is what PathStyles receives.
type PathStylesServiceRequestOptions struct {
	PathParams *PathStylesPathParams
	RawRequest *http.Request
}

// Validate checks the parameters and the body against the constraints of the spec.
func (o *PathStylesServiceRequestOptions) Validate() error {
	var errs runtime.ValidationErrors
	if o.PathParams != nil {
		errs.Append("path", o.PathParams.Validate())
	}
	return errs.Err()
}

// PathStylesResponseData is what PathStyles returns.
type PathStylesResponseData struct {
	Status  int
	Headers http.Header
	Body    any

	contentType string
}

// NewPathStylesResponseData returns the 200 response with its application/json body.
func NewPathStylesResponseData(body Echo) *PathStylesResponseData {
	return &PathStylesResponseData{Status: 200, Body: body, contentType: "application/json"}
}

// WithStatus sets the status code.
func (r *PathStylesResponseData) WithStatus(code int) *PathStylesResponseData {
	r.Status = code
	return r
}

// WithHeaders sets the headers.
func (r *PathStylesResponseData) WithHeaders(h http.Header) *PathStylesResponseData {
	r.Headers = h
	return r
}

// StatusCode returns the status.
func (r *PathStylesResponseData) StatusCode() int {
	return r.Status
}

// Header returns the headers.
func (r *PathStylesResponseData) Header() http.Header {
	return r.Headers
}

// Payload returns the body.
func (r *PathStylesResponseData) Payload() any {
	return r.Body
}

// ContentType returns the media type of the body, empty for the default of its Go type.
func (r *PathStylesResponseData) ContentType() string {
	return r.contentType
}

// QueryStylesServiceRequestOptions is what QueryStyles receives.
type QueryStylesServiceRequestOptions struct {
	Query      *QueryStylesQuery
	RawRequest *http.Request
}

// Validate checks the parameters and the body against the constraints of the spec.
func (o *QueryStylesServiceRequestOptions) Validate() error {
	return nil
}

// QueryStylesResponseData is what QueryStyles returns.
type QueryStylesResponseData struct {
	Status  int
	Headers http.Header
	Body    any

	contentType string
}

// NewQueryStylesResponseData returns the 200 response with its application/json body.
func NewQueryStylesResponseData(body Echo) *QueryStylesResponseData {
	return &QueryStylesResponseData{Status: 200, Body: body, contentType: "application/json"}
}

// WithStatus sets the status code.
func (r *QueryStylesResponseData) WithStatus(code int) *QueryStylesResponseData {
	r.Status = code
	return r
}

// WithHeaders sets the headers.
func (r *QueryStylesResponseData) WithHeaders(h http.Header) *QueryStylesResponseData {
	r.Headers = h
	return r
}

// StatusCode returns the status.
func (r *QueryStylesResponseData) StatusCode() int {
	return r.Status
}

// Header returns the headers.
func (r *QueryStylesResponseData) Header() http.Header {
	return r.Headers
}

// Payload returns the body.
func (r *QueryStylesResponseData) Payload() any {
	return r.Body
}

// ContentType returns the media type of the body, empty for the default of its Go type.
func (r *QueryStylesResponseData) ContentType() string {
	return r.contentType
}

// HeaderStylesServiceRequestOptions is what HeaderStyles receives.
type HeaderStylesServiceRequestOptions struct {
	Headers    *HeaderStylesHeaders
	RawRequest *http.Request
}

// Validate checks the parameters and the body against the constraints of the spec.
func (o *HeaderStylesServiceRequestOptions) Validate() error {
	return nil
}

// HeaderStylesResponseData is what HeaderStyles returns.
type HeaderStylesResponseData struct {
	Status  int
	Headers http.Header
	Body    any

	contentType string
}

// NewHeaderStylesResponseData returns the 200 response with its application/json body.
func NewHeaderStylesResponseData(body Echo) *HeaderStylesResponseData {
	return &HeaderStylesResponseData{Status: 200, Body: body, contentType: "application/json"}
}

// WithStatus sets the status code.
func (r *HeaderStylesResponseData) WithStatus(code int) *HeaderStylesResponseData {
	r.Status = code
	return r
}

// WithHeaders sets the headers.
func (r *HeaderStylesResponseData) WithHeaders(h http.Header) *HeaderStylesResponseData {
	r.Headers = h
	return r
}

// StatusCode returns the status.
func (r *HeaderStylesResponseData) StatusCode() int {
	return r.Status
}

// Header returns the headers.
func (r *HeaderStylesResponseData) Header() http.Header {
	return r.Headers
}

// Payload returns the body.
func (r *HeaderStylesResponseData) Payload() any {
	return r.Body
}

// ContentType returns the media type of the body, empty for the default of its Go type.
func (r *HeaderStylesResponseData) ContentType() string {
	return r.contentType
}

// CookieStylesServiceRequestOptions is what CookieStyles receives.
type CookieStylesServiceRequestOptions struct {
	Cookies    *CookieStylesCookies
	RawRequest *http.Request
}

// Validate checks the parameters and the body against the constraints of the spec.
func (o *CookieStylesServiceRequestOptions) Validate() error {
	return nil
}

// CookieStylesResponseData is what CookieStyles returns.
type CookieStylesResponseData struct {
	Status  int
	Headers http.Header
	Body    any

	contentType string
}

// NewCookieStylesResponseData returns the 200 response with its application/json body.
func NewCookieStylesResponseData(body Echo) *CookieStylesResponseData {
	return &CookieStylesResponseData{Status: 200, Body: body, contentType: "application/json"}
}

// WithStatus sets the status code.
func (r *CookieStylesResponseData) WithStatus(code int) *CookieStylesResponseData {
	r.Status = code
	return r
}

// WithHeaders sets the headers.
func (r *CookieStylesResponseData) WithHeaders(h http.Header) *CookieStylesResponseData {
	r.Headers = h
	return r
}

// StatusCode returns the status.
func (r *CookieStylesResponseData) StatusCode() int {
	return r.Status
}

// Header returns the headers.
func (r *CookieStylesResponseData) Header() http.Header {
	return r.Headers
}

// Payload returns the body.
func (r *CookieStylesResponseData) Payload() any {
	return r.Body
}

// ContentType returns the media type of the body, empty for the default of its Go type.
func (r *CookieStylesResponseData) ContentType() string {
	return r.contentType
}

// The error types the handlers use, as the runtime declares them.
type (
	ErrorKind           = runtime.ErrorKind
	HandlerError        = runtime.HandlerError
	ErrorHandler        = runtime.ErrorHandler
	ErrorHandlerFunc    = runtime.ErrorHandlerFunc
	DefaultErrorHandler = runtime.DefaultErrorHandler
)

// The kinds of HandlerError.
const (
	ErrorParse      = runtime.ErrorParse
	ErrorDecode     = runtime.ErrorDecode
	ErrorValidation = runtime.ErrorValidation
	ErrorService    = runtime.ErrorService
	ErrorResponse   = runtime.ErrorResponse
)

// ServerOptions is what the adapter and the router are set up with.
type ServerOptions struct {
	Router             any
	Middleware         []func(http.Handler) http.Handler
	ErrorHandler       runtime.ErrorHandler
	JSONDecoder        func(body io.Reader, dst any, isRequired bool) error
	MultipartMaxMemory int64
}

// ServerOption sets one field of ServerOptions.
type ServerOption func(*ServerOptions)

// NewServerOptions applies opts to the defaults.
func NewServerOptions(opts ...ServerOption) *ServerOptions {
	o := &ServerOptions{
		ErrorHandler:       runtime.DefaultErrorHandler{},
		JSONDecoder:        runtime.DecodeJSON,
		MultipartMaxMemory: 33554432,
	}
	for _, opt := range opts {
		opt(o)
	}
	return o
}

// WithMiddleware wraps the routes with mw, outermost first; on a new router, unknown paths too.
func WithMiddleware(mw ...func(http.Handler) http.Handler) ServerOption {
	return func(o *ServerOptions) {
		o.Middleware = append(o.Middleware, mw...)
	}
}

// WithErrorHandler sets what writes the response of a failed request.
func WithErrorHandler(h runtime.ErrorHandler) ServerOption {
	return func(o *ServerOptions) {
		o.ErrorHandler = h
	}
}

// WithJSONDecoder sets what reads JSON bodies. isRequired says whether an empty body is an error.
func WithJSONDecoder(decode func(body io.Reader, dst any, isRequired bool) error) ServerOption {
	return func(o *ServerOptions) {
		o.JSONDecoder = decode
	}
}

// WithMultipartMaxMemory sets how much of a multipart form stays in memory.
func WithMultipartMaxMemory(n int64) ServerOption {
	return func(o *ServerOptions) {
		o.MultipartMaxMemory = n
	}
}

// HTTPAdapter answers HTTP requests by calling the service.
type HTTPAdapter struct {
	svc  ServiceInterface
	opts *ServerOptions
}

type responseData interface {
	StatusCode() int
	Header() http.Header
	Payload() any
	ContentType() string
}

// NewHTTPAdapter returns the adapter of svc.
func NewHTTPAdapter(svc ServiceInterface, opts ...ServerOption) *HTTPAdapter {
	return &HTTPAdapter{svc: svc, opts: NewServerOptions(opts...)}
}

// PathStyles handles GET /path/{simple}/{label}/{matrix}/{list}.
func (a *HTTPAdapter) PathStyles(w http.ResponseWriter, r *http.Request) {
	opts := &PathStylesServiceRequestOptions{RawRequest: r}
	opts.PathParams = &PathStylesPathParams{}
	if err := runtime.DecodePath(r.PathValue("simple"), runtime.Param{Name: "simple", Style: runtime.StyleSimple, IsExplode: false, IsRequired: true, IsJSON: false}, &opts.PathParams.Simple); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "PathStyles", ParamName: "simple", ParamLocation: "path", Err: err})
		return
	}
	if err := runtime.DecodePath(r.PathValue("label"), runtime.Param{Name: "label", Style: runtime.StyleLabel, IsExplode: false, IsRequired: true, IsJSON: false}, &opts.PathParams.Label); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "PathStyles", ParamName: "label", ParamLocation: "path", Err: err})
		return
	}
	if err := runtime.DecodePath(r.PathValue("matrix"), runtime.Param{Name: "matrix", Style: runtime.StyleMatrix, IsExplode: false, IsRequired: true, IsJSON: false}, &opts.PathParams.Matrix); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "PathStyles", ParamName: "matrix", ParamLocation: "path", Err: err})
		return
	}
	if err := runtime.DecodePath(r.PathValue("list"), runtime.Param{Name: "list", Style: runtime.StyleSimple, IsExplode: true, IsRequired: true, IsJSON: false}, &opts.PathParams.List); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "PathStyles", ParamName: "list", ParamLocation: "path", Err: err})
		return
	}

	res, err := a.svc.PathStyles(r.Context(), opts)
	if err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "PathStyles", Err: err})
		return
	}
	if res == nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "PathStyles", Err: runtime.ErrNoResponse})
		return
	}
	a.write(w, r, "PathStyles", res)
}

// QueryStyles handles GET /query.
func (a *HTTPAdapter) QueryStyles(w http.ResponseWriter, r *http.Request) {
	opts := &QueryStylesServiceRequestOptions{RawRequest: r}
	query := r.URL.Query()
	opts.Query = &QueryStylesQuery{}
	if err := runtime.DecodeQuery(query, runtime.Param{Name: "form", Style: runtime.StyleForm, IsExplode: true, IsRequired: false, IsJSON: false}, &opts.Query.Form); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "QueryStyles", ParamName: "form", ParamLocation: "query", Err: err})
		return
	}
	if err := runtime.DecodeQuery(query, runtime.Param{Name: "csv", Style: runtime.StyleForm, IsExplode: false, IsRequired: false, IsJSON: false}, &opts.Query.Csv); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "QueryStyles", ParamName: "csv", ParamLocation: "query", Err: err})
		return
	}
	if err := runtime.DecodeQuery(query, runtime.Param{Name: "space", Style: runtime.StyleSpaceDelimited, IsExplode: false, IsRequired: false, IsJSON: false}, &opts.Query.Space); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "QueryStyles", ParamName: "space", ParamLocation: "query", Err: err})
		return
	}
	if err := runtime.DecodeQuery(query, runtime.Param{Name: "pipe", Style: runtime.StylePipeDelimited, IsExplode: false, IsRequired: false, IsJSON: false}, &opts.Query.Pipe); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "QueryStyles", ParamName: "pipe", ParamLocation: "query", Err: err})
		return
	}
	if err := runtime.DecodeQuery(query, runtime.Param{Name: "deep", Style: runtime.StyleDeepObject, IsExplode: false, IsRequired: false, IsJSON: false}, &opts.Query.Deep); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "QueryStyles", ParamName: "deep", ParamLocation: "query", Err: err})
		return
	}
	if err := runtime.DecodeQuery(query, runtime.Param{Name: "flat", Style: runtime.StyleForm, IsExplode: false, IsRequired: false, IsJSON: false}, &opts.Query.Flat); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "QueryStyles", ParamName: "flat", ParamLocation: "query", Err: err})
		return
	}
	if err := runtime.DecodeQuery(query, runtime.Param{Name: "json", Style: runtime.StyleForm, IsExplode: true, IsRequired: false, IsJSON: true}, &opts.Query.JSON); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "QueryStyles", ParamName: "json", ParamLocation: "query", Err: err})
		return
	}
	if err := runtime.DecodeQuery(query, runtime.Param{Name: "needed", Style: runtime.StyleForm, IsExplode: true, IsRequired: true, IsJSON: false}, &opts.Query.Needed); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "QueryStyles", ParamName: "needed", ParamLocation: "query", Err: err})
		return
	}

	res, err := a.svc.QueryStyles(r.Context(), opts)
	if err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "QueryStyles", Err: err})
		return
	}
	if res == nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "QueryStyles", Err: runtime.ErrNoResponse})
		return
	}
	a.write(w, r, "QueryStyles", res)
}

// HeaderStyles handles GET /header.
func (a *HTTPAdapter) HeaderStyles(w http.ResponseWriter, r *http.Request) {
	opts := &HeaderStylesServiceRequestOptions{RawRequest: r}
	opts.Headers = &HeaderStylesHeaders{}
	if err := runtime.DecodeHeader(r.Header, runtime.Param{Name: "X-Tags", Style: runtime.StyleSimple, IsExplode: false, IsRequired: false, IsJSON: false}, &opts.Headers.XTags); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "HeaderStyles", ParamName: "X-Tags", ParamLocation: "header", Err: err})
		return
	}
	if err := runtime.DecodeHeader(r.Header, runtime.Param{Name: "X-Point", Style: runtime.StyleSimple, IsExplode: false, IsRequired: false, IsJSON: false}, &opts.Headers.XPoint); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "HeaderStyles", ParamName: "X-Point", ParamLocation: "header", Err: err})
		return
	}
	if err := runtime.DecodeHeader(r.Header, runtime.Param{Name: "X-When", Style: runtime.StyleSimple, IsExplode: false, IsRequired: false, IsJSON: false}, &opts.Headers.XWhen); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "HeaderStyles", ParamName: "X-When", ParamLocation: "header", Err: err})
		return
	}

	res, err := a.svc.HeaderStyles(r.Context(), opts)
	if err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "HeaderStyles", Err: err})
		return
	}
	if res == nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "HeaderStyles", Err: runtime.ErrNoResponse})
		return
	}
	a.write(w, r, "HeaderStyles", res)
}

// CookieStyles handles GET /cookie.
func (a *HTTPAdapter) CookieStyles(w http.ResponseWriter, r *http.Request) {
	opts := &CookieStylesServiceRequestOptions{RawRequest: r}
	opts.Cookies = &CookieStylesCookies{}
	if err := runtime.DecodeCookie(r.Cookies(), runtime.Param{Name: "session", Style: runtime.StyleForm, IsExplode: true, IsRequired: false, IsJSON: false}, &opts.Cookies.Session); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "CookieStyles", ParamName: "session", ParamLocation: "cookie", Err: err})
		return
	}
	if err := runtime.DecodeCookie(r.Cookies(), runtime.Param{Name: "flags", Style: runtime.StyleForm, IsExplode: false, IsRequired: false, IsJSON: false}, &opts.Cookies.Flags); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "CookieStyles", ParamName: "flags", ParamLocation: "cookie", Err: err})
		return
	}

	res, err := a.svc.CookieStyles(r.Context(), opts)
	if err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "CookieStyles", Err: err})
		return
	}
	if res == nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "CookieStyles", Err: runtime.ErrNoResponse})
		return
	}
	a.write(w, r, "CookieStyles", res)
}

func (a *HTTPAdapter) fail(w http.ResponseWriter, r *http.Request, err *runtime.HandlerError) {
	// A response that failed to write leaves its media type, which is not the error's.
	w.Header().Del("Content-Type")
	a.opts.ErrorHandler.HandleError(w, r, err.StatusCode(), err)
}

func (a *HTTPAdapter) failDecode(w http.ResponseWriter, r *http.Request, id string, err error) {
	a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorDecode, OperationID: id, Err: err})
}

func (a *HTTPAdapter) write(w http.ResponseWriter, r *http.Request, id string, res responseData) {
	if res.ContentType() != "" {
		w.Header().Set("Content-Type", res.ContentType())
	}
	err := runtime.Write(w, res.StatusCode(), res.Header(), res.Payload())
	switch {
	case errors.Is(err, runtime.ErrContentType):
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorResponse, OperationID: id, Err: err})
	case err != nil:
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: id, Err: err})
	}
}

// WithRouter registers the routes on r instead of a new ControllerRegister.
func WithRouter(r *web.ControllerRegister) ServerOption {
	return func(o *ServerOptions) {
		o.Router = r
	}
}

// NewRouter registers every operation on a beego ControllerRegister.
func NewRouter(svc ServiceInterface, opts ...ServerOption) *web.ControllerRegister {
	o := NewServerOptions(opts...)
	a := NewHTTPAdapter(svc, opts...)
	wrap := func(h http.Handler) http.Handler {
		for i := len(o.Middleware) - 1; i >= 0; i-- {
			h = o.Middleware[i](h)
		}
		return h
	}
	register := func(r *web.ControllerRegister, route func(http.Handler) http.Handler) {
		r.AddMethod("GET", "/path/:simple/:label/:matrix/:list", handle(route(http.HandlerFunc(a.PathStyles))))
		r.AddMethod("GET", "/query", handle(route(http.HandlerFunc(a.QueryStyles))))
		r.AddMethod("GET", "/header", handle(route(http.HandlerFunc(a.HeaderStyles))))
		r.AddMethod("GET", "/cookie", handle(route(http.HandlerFunc(a.CookieStyles))))
	}

	if r, _ := o.Router.(*web.ControllerRegister); r != nil {
		register(r, wrap)
		return r
	}
	r := web.NewControllerRegister()
	register(r, wrap)
	r.Any("/*", handle(wrap(http.NotFoundHandler())))
	return r
}

func handle(h http.Handler) web.HandleFunc {
	return func(c *bcontext.Context) {
		r := c.Request
		for key, value := range c.Input.Params() {
			if len(key) > 1 && key[0] == ':' {
				r.SetPathValue(key[1:], value)
			}
		}
		// beego has read a form body already, so it goes back into the body for h.
		if r.PostForm != nil && strings.HasPrefix(r.Header.Get("Content-Type"), "application/x-www-form-urlencoded") {
			r.Body = io.NopCloser(strings.NewReader(r.PostForm.Encode()))
		}
		h.ServeHTTP(c.ResponseWriter, r)
	}
}